Refactor Path utility class
