+++ src/integrationTest/java/com/demo/filmdb/ApiRootIntegrationTests.java
@@ -8,8 +8,8 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.web.context.WebApplicationContext;
 
-import static com.demo.filmdb.Utils.API_PREFIX;
 import static com.demo.filmdb.Utils.configureMockMvc;
+import static com.demo.filmdb.util.Path.API_PREFIX;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
