+++ src/main/java/com/demo/filmdb/RootController.java
@@ -14,9 +14,9 @@ import org.springframework.web.bind.annotation.RestController;
 import java.util.HashMap;
 import java.util.Map;
 
-import static com.demo.filmdb.utils.Path.API_PREFIX;
 import static com.demo.filmdb.config.SpringDocConfig.SUCCESS;
 import static com.demo.filmdb.config.SpringDocConfig.TAG_ROOT;
+import static com.demo.filmdb.util.Path.API_PREFIX;
 import static org.springframework.hateoas.server.core.DummyInvocationUtils.methodOn;
 import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.linkTo;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
