+++ src/main/java/com/demo/filmdb/security/JwtAuthController.java
@@ -17,9 +17,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import static com.demo.filmdb.utils.Path.API_PREFIX;
-import static com.demo.filmdb.utils.Path.LOGIN;
 import static com.demo.filmdb.config.SpringDocConfig.TAG_LOGIN;
+import static com.demo.filmdb.util.Path.API_PREFIX;
+import static com.demo.filmdb.util.Path.LOGIN;
 import static org.springframework.http.HttpHeaders.WWW_AUTHENTICATE;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
