+++ src/main/java/com/demo/filmdb/util/Path.java
@@ -1,0 +1,6 @@
+package com.demo.filmdb.util;
+
+public class Path {
+    public static final String API_PREFIX = "/api";
+    public static final String LOGIN = "/login";
+}
