--- src/main/java/com/demo/filmdb/utils/Path.java
@@ -1,9 +1,0 @@
-package com.demo.filmdb.utils;
-
-public class Path {
-    public static final String API_PREFIX = "/api";
-
-    public static final String LOGIN = "/login";
-    public static final String FILM = "/films";
-    public static final String PEOPLE = "/people";
-}
